Improve horizontal scrolling for project page view
